requirements.txt
icalevents
icalendar
requests
python-dateutil
